Start removing static utility type methods into separate module.
