+++ pywrds/wrdsapi.py
@@ -13,6 +13,7 @@ from _wrds_db_descriptors import WRDS_DOMAIN, _GET_ALL, FIRST_DATES, \
     FIRST_DATE_GUESSES, AUTOEXEC_TEXT
 
 from pywrds import sshlib
+from pywrds import utility as wrds_util
 
 
 class WrdsSession(object):
@@ -79,29 +80,6 @@ class WrdsSession(object):
         except:
             raise Warning("Need to implement login without key authentication")
 
-    @staticmethod
-    def rows_per_file_adjusted(dataset):
-        """rows_per_file_adjusted(dataset)
-
-        _rows_per_file chooses a number of rows to query in each
-        _get_wrds_chunk request to ensure that the files produced
-        do not approach the 1 GB server limit.  For most datasets,
-        10^7 rows in a file is not a problem.  For optionm.opprcd,
-        this number is dropped to 10^6.
-
-        To date optionm.opprcd is the only dataset for which this has
-        consistently been necessary.  This is subject to change
-        with further use cases.
-
-        return rows_per_file
-        """
-        rows_per_file = 10**7
-        if dataset.replace('.', '_') == 'optionm_opprcd':
-            rows_per_file = 10**6
-        elif dataset.replace('.', '_') == 'optionm_optionmnames':
-            rows_per_file = 10**6
-        return rows_per_file
-
     def get_ymd_range(self, min_date, dataset, weekdays=1):
         """get_ymd_range(min_date, dataset, weekdays=1) gets a list of
         tuples [year, month, date] over which to iterate in wrds_loop.  Some
@@ -116,7 +94,7 @@ class WrdsSession(object):
         ymdrange = []
         years = xrange(min_year, self.now.tm_year+1)
         for year in years:
-            frequency = self.get_loop_frequency(dataset, year)
+            frequency = wrds_util.get_loop_frequency(dataset, year)
             if frequency == 'Y':
                 new_ymd = [year, 0, 0]
                 ymdrange.append(new_ymd)
@@ -126,106 +104,13 @@ class WrdsSession(object):
             elif frequency == 'D':
                 new_ymd = [[year, x, y] for x in range(1, 13) for y in range(
                     1, 32)]
-                new_ymd = self.fix_weekdays(new_ymd, weekdays)
+                new_ymd = wrds_util.fix_weekdays(new_ymd, weekdays)
                 ymdrange = ymdrange + new_ymd
 
         ymdrange = [x for x in ymdrange if x <= [self.this_year,
                                                  self.this_month, self.today]]
         ymdrange = [x for x in ymdrange if x >= [min_year, min_month, min_day]]
         return ymdrange
-
-    @staticmethod
-    def get_loop_frequency(dataset, year):
-        """get_loop_frequency(dataset, year) finds the best frequency at which
-        to query the server for the given dataset so as to avoid producing
-        problematically large files.
-
-        return frequency
-        """
-        frequency = 'Y'
-        if dataset.startswith('optionm'):
-            if year < 2008:
-                frequency = 'M'
-            else:
-                frequency = 'D'
-        elif re.search('det_xepsus', dataset, flags=re.I):
-            if year > 2005:
-                frequency = 'M'
-        elif re.search('det_xepsint', dataset, flags=re.I):
-            if year > 2003:
-                frequency = 'M'
-        elif re.search('taq', dataset, flags=re.I):
-            frequency = 'D'
-        return frequency
-
-    @staticmethod
-    def fix_weekdays(ymds, weekdays=1):
-        """fix_weekdays(ymds, weekdays=1) takes a set of [year,month,date]
-        tuples "ymds" and removes those which are not valid days,
-        e.g. June 31, February 30.
-
-        If weekdays is set to its default value of 1, it also removes
-        Saturday and Sundays.
-
-        return ymds
-        """
-        ymds2 = []
-        for [y, m, d] in ymds:
-            try:
-                wday = datetime.date(y, m, d).weekday()
-            except ValueError:
-                wday = -1
-            if weekdays == 1 and wday in range(5):
-                # weekdays==1 --> only keep weekdays     #
-                ymds2.append([y, m, d])
-            elif weekdays == 0 and wday != -1:
-                # weekdays==0 --> keey any valid day     #
-                ymds2.append([y, m, d])
-        return ymds2
-
-    @staticmethod
-    def fix_input_name(dataset, year, month, day, rows=[]):
-        """fix_input_name(dataset, year, month, day, rows=[])
-        adjusts the user-supplied dataset name to use the same
-        upper/lower case conventions as WRDS does.
-
-        return [dataset, output_file]
-        """
-        [Y, M, D, R] = [year, month, day, rows]
-        if year != 'all':
-            ystr = '_'*(dataset[-1].isdigit()) + str(Y)
-            mstr = '' + (M != 0)*('0'*(month < 10) + str(M))
-            dstr = (D != 0)*('0'*(D < 10) + str(D))
-            ymdstr = ystr + mstr + dstr + '.tsv'
-            output_file = re.sub('\.', '_', dataset) + ymdstr
-        else:
-            output_file = re.sub('\.', '_', dataset) + '.tsv'
-
-        if dataset.lower() == 'optionm.opprcd':
-            dataset = dataset + str(year)
-
-        elif dataset.lower() in ['taq.cq', 'taq.ct']:
-            dataset = re.sub('cq', 'CQ', dataset)
-            dataset = re.sub('ct', 'CT', dataset)
-            ystr = '_' + str(Y)
-            mstr = '' + (M != 0)*('0'*(M < 10) + str(M))
-            dstr = '' + (D != 0)*('0'*(D<10) + str(D))
-            ymdstr = ystr + mstr + dstr
-            dataset = dataset + ymdstr
-
-        elif dataset.lower() in ['taq.mast', 'taq.div']:
-            ymdstr = '_' + str(Y) + (M != 0)*('0'*(M < 10) + str(M))
-            dataset = dataset + ymdstr
-
-        elif dataset.lower() == 'taq.rgsh':
-            ymdstr = str(100*Y + M)[2:]
-            dataset = 'taq.RGSH' + ymdstr
-
-        if R != []:
-            rowstr = 'rows' + str(R[0]) + 'to' + str(R[1]) + '.tsv'
-            output_file = re.sub('.tsv$', '', output_file) + rowstr
-
-        return [dataset, output_file]
 
     def wrds_sas_script(self, dataset, year, month=0, day=0, rows=[]):
         """wrds_sas_script(dataset, year, month=0, day=0, rows=[])
@@ -248,23 +133,23 @@ class WrdsSession(object):
             sas_file = sas_file + ymdstr
         sas_file = sas_file + '.sas'
 
-        [dataset, output_file] = self.fix_input_name(dataset, Y, M, D, R)
+        [dataset, output_file] = wrds_util.fix_input_name(dataset, Y, M, D, R)
         with open(os.path.join(self.download_path, sas_file), 'wb') as fd:
             fd.write('DATA new_data;\n')
             fd.write('\tSET ' + dataset)
             if Y != 'all':
                 where_query = ' (where = ('
-                year_query = ('(year(' + self.wrds_datevar(dataset) + ')'
+                year_query = ('(year(' + wrds_util.wrds_datevar(dataset) + ')'
                     + ' between ' + str(Y) + ' and ' + str(Y) + ')')
                 where_query = where_query + year_query
 
                 if M != 0:
-                    month_query = (' and (month(' + self.wrds_datevar(dataset)
+                    month_query = (' and (month(' + wrds_util.wrds_datevar(dataset)
                         +') between '+str(M)+' and '+str(M)+')')
                     where_query = where_query + month_query
 
                 if D != 0:
-                    day_query = (' and (day(' + self.wrds_datevar(dataset)
+                    day_query = (' and (day(' + wrds_util.wrds_datevar(dataset)
                         +') between '+str(D)+' and '+str(D)+')')
                     where_query = where_query+day_query
 
@@ -404,30 +289,6 @@ class WrdsSession(object):
 
         return [min_year, min_month, min_day]
 
-    @staticmethod
-    def wrds_datevar(filename):
-        """wrds_datevar(filename)
-        Different datasets in WRDS use different names for
-        their date-variables.  wrds_datevar gives the right date
-        variable for each dataset.  This may need periodic updating.
-        Crowdsourcing is welcome.
-
-        return date_var
-        """
-        if filename in ['tfn.s12', 'tfn.s34']:
-            return 'fdate'
-        if re.search('^crsp', filename):
-            return 'date'
-        if re.search('^comp', filename):
-            return 'DATADATE'
-        if re.search('^optionm\.opprcd', filename):
-            return 'date'
-        if re.search('^optionm', filename):
-            return 'effect_date'
-        if re.search('^ibes', filename):
-            return 'anndats'
-        return 'date'
-
     def setup_wrds_key(self):
         """setup_wrds_key() sets up a key-based authentication on
         the wrds server, so that the user can log in without a
@@ -502,15 +363,15 @@ class WrdsSession(object):
         tic = time.time()
         keep_going = 1
         [startrow, numfiles, total_rows, tic] = [1, 0, 0, time.time()]
-        rows_per_file = self.rows_per_file_adjusted(dataset)
-        [dset2, outfile] = self.fix_input_name(dataset, Y, M, D, [])
+        rows_per_file = wrds_util.rows_per_file_adjusted(dataset)
+        [dset2, outfile] = wrds_util.fix_input_name(dataset, Y, M, D, [])
 
         # Check if output file in local dir, if not send request.
         if os.path.exists(os.path.join(self.download_path, outfile)):
             keep_going = 0
         while keep_going:
             R = [startrow, startrow - 1 + rows_per_file]
-            [dset2, outfile] = self.fix_input_name(dataset, Y, M, D, R)
+            [dset2, outfile] = wrds_util.fix_input_name(dataset, Y, M, D, R)
 
             if not os.path.exists(os.path.join(self.download_path, outfile)):
                 [keep_going, dt] = self._get_wrds_chunk(dataset, Y, M, D, R)
@@ -520,7 +381,7 @@ class WrdsSession(object):
                 if os.path.exists(os.path.join(self.download_path, outfile)):
                     log_lines = self.get_numlines_from_log(
                         outfile, dname=self.download_path)
-                    numlines = self.get_numlines(os.path.join(
+                    numlines = wrds_util.get_numlines(os.path.join(
                         self.download_path, outfile))
                     if log_lines > numlines:
                         print('get_wrds error: file '
@@ -600,7 +461,7 @@ class WrdsSession(object):
                 remote_size = self._wait_for_sas_file_completion(outfile)
                 [get_success, dt] = self._retrieve_file(outfile, remote_size)
                 local_size \
-                    = self._wait_for_retrieve_completion(outfile, get_success)
+                    = wrds_util._wait_for_retrieve_completion(outfile, get_success)
                 compare_success = \
                     self._compare_local_to_remote(outfile, remote_size,
                                                   local_size)
@@ -703,7 +564,7 @@ class WrdsSession(object):
 
         for ymd in self.get_ymd_range(min_date, dataset, 1):
             [Y, M, D] = ymd
-            [dset2, outfile] = self.fix_input_name(dataset, Y, M, D, [])
+            [dset2, outfile] = wrds_util.fix_input_name(dataset, Y, M, D, [])
             if outfile in flist:
                 continue
             get_output = self.get_wrds(dataset, Y, M=M, D=D, recombine=recombine)
@@ -980,37 +841,6 @@ class WrdsSession(object):
 
         return [get_success, time.time()-tic]
 
-    @staticmethod
-    def _wait_for_retrieve_completion(outfile, get_success, maxwait=1200):
-        """_wait_for_retrieve_completion(outfile, get_success) checks the
-        size of the downloaded file outfile multiple times and waits for
-        two successive measurements giving the same file size.  Until
-        this point, it infers that the download is still in progress.
-
-        return local_size
-        """
-        if get_success == 0:
-            return 0
-        waited3 = 0
-        [locmeasure1, locmeasure2, mtime2] = [0, 1, time.time()]
-        write_file = '.' + outfile + '--writing'
-        local_path = os.path.join(os.path.expanduser('~'), write_file)
-        while ((waited3 < maxwait)
-            and (locmeasure1 != locmeasure2 or (time.time() - mtime2) <= 10)):
-            locmeasure1 = locmeasure2
-            time.sleep(5)
-            waited3 += 5
-            local_stat = os.stat(local_path)
-            locmeasure2 = local_stat.st_size
-            mtime2 = local_stat.st_mtime
-
-        if waited3 >= maxwait:
-            print('get_wrds stopped waiting for SAS completion at step 3',
-                locmeasure1, locmeasure2, mtime2)
-            locmeasure1 = 0
-        local_size = locmeasure1
-        return local_size
-
     def _compare_local_to_remote(self, outfile, remote_size, local_size):
         """_compare_local_to_remote(outfile, remote_size, local_size)
         compares the size of the file "outfile" downloaded (local_size) to
@@ -1158,7 +988,7 @@ class WrdsSession(object):
         if os.path.exists(os.path.join(dname, fname + '.tsv')):
             isready = 0
 
-        rows_per_file = self.rows_per_file_adjusted(fname0)
+        rows_per_file = wrds_util.rows_per_file_adjusted(fname0)
         flist0 = os.listdir(dname)
         flist0 = [x for x in flist0 if x.endswith('.tsv')]
         flist0 = [x for x in flist0 if re.search(fname0, x)]
@@ -1192,7 +1022,7 @@ class WrdsSession(object):
                                                            + '.tsv')]
             if len(flist2) == 1:
                 outfile = flist2[0]
-                numlines = self.get_numlines(os.path.join(dname, outfile))
+                numlines = wrds_util.get_numlines(os.path.join(dname, outfile))
                 log_numlines = self.get_numlines_from_log(outfile, dname)
                 if numlines != log_numlines:
                     isready = 0
@@ -1222,7 +1052,7 @@ class WrdsSession(object):
             return combined_files
 
         fname0 = re.sub('rows[0-9][0-9]*to[0-9][0-9]*\.tsv', '', fname)
-        rows_per_file = self.rows_per_file_adjusted(fname0)
+        rows_per_file = wrds_util.rows_per_file_adjusted(fname0)
 
         flist0 = [x for x in os.listdir(dname) if re.search(fname0, x)]
         flist0 = [x for x in flist0 if x.endswith('.tsv')]
@@ -1274,22 +1104,6 @@ class WrdsSession(object):
             for fname1 in flist:
                 os.remove(os.path.join(dname, fname1))
         return combined_files
-
-    @staticmethod
-    def get_numlines(path2file):
-        """get_numlines(path2file) reads a textfile located at
-        path2file and returns the number of lines found.
-
-        return numlines
-        """
-        with open(path2file, 'rb') as fd:
-            fsize = os.stat(fd.name).st_size
-            numlines = 0
-            first_line = fd.readline().split('\t')
-            while fd.tell() < fsize:
-                fline = fd.readline()
-                numlines += 1
-        return numlines
 
     def _try_put(self, local_path, remote_path, domain, username, ports=[22]):
         """_try_put(local_path, remote_path, domain, username, ports)
